refactor(nutrition): extract page loading helpers
